+++ hp3parclient/exceptions.py
@@ -1,6 +1,6 @@
 # vim: tabstop=4 shiftwidth=4 softtabstop=4
 #
-# Copyright 2012 Hewlett Packard Development Company, L.P.
+# Copyright 2012-2016 Hewlett Packard Development Company, L.P.
 # All Rights Reserved.
 #
 #    Licensed under the Apache License, Version 2.0 (the "License"); you may
@@ -25,6 +25,12 @@ Exceptions for the client
 """
 
 import logging
+
+# Python 3+ override
+try:
+    basestring
+except NameError:
+    basestring = str
 
 LOG = logging.getLogger(__name__)
 
@@ -69,7 +75,7 @@ class ClientException(Exception):
         if not error:
             return
 
-        if isinstance(error, str):
+        if isinstance(error, basestring):
             # instead of KeyError below, take it and make it the _error_desc.
             self._error_desc = error
         else: